--- a/modules/imgproc/src/color_lab.cpp
+++ b/modules/imgproc/src/color_lab.cpp
@@ -1609,30 +1609,36 @@ struct RGB2Lab_b
             }
 
 #define RGB2LAB_BATCH(n) \
-            v_int32x4 vR##n(tab[vRi.val[4*n+0]], tab[vRi.val[4*n+1]], tab[vRi.val[4*n+2]], tab[vRi.val[4*n+3]]); \
-            v_int32x4 vG##n(tab[vGi.val[4*n+0]], tab[vGi.val[4*n+1]], tab[vGi.val[4*n+2]], tab[vGi.val[4*n+3]]); \
-            v_int32x4 vB##n(tab[vBi.val[4*n+0]], tab[vBi.val[4*n+1]], tab[vBi.val[4*n+2]], tab[vBi.val[4*n+3]]); \
+            v_int32x4 vR##n(tab[v_extract_n<4*n+0>(vRi)], tab[v_extract_n<4*n+1>(vRi)], \
+                            tab[v_extract_n<4*n+2>(vRi)], tab[v_extract_n<4*n+3>(vRi)]); \
+            v_int32x4 vG##n(tab[v_extract_n<4*n+0>(vGi)], tab[v_extract_n<4*n+1>(vGi)], \
+                            tab[v_extract_n<4*n+2>(vGi)], tab[v_extract_n<4*n+3>(vGi)]); \
+            v_int32x4 vB##n(tab[v_extract_n<4*n+0>(vBi)], tab[v_extract_n<4*n+1>(vBi)], \
+                            tab[v_extract_n<4*n+2>(vBi)], tab[v_extract_n<4*n+3>(vBi)]); \
             \
             /* int fX = LabCbrtTab_b[CV_DESCALE(R*C0 + G*C1 + B*C2, lab_shift)];*/ \
             v_int32x4 vfX##n = v_fma(vR##n, v_setall_s32(C0), v_setall_s32(one_lsh_by_lab_shift_sub_one)); \
             vfX##n = v_fma(vG##n, v_setall_s32(C1), vfX##n); \
             vfX##n = v_fma(vB##n, v_setall_s32(C2), vfX##n); \
             vfX##n = v_shr<lab_shift>(vfX##n); \
-            vfX##n = v_int32x4(LabCbrtTab_b[vfX##n.val[0]], LabCbrtTab_b[vfX##n.val[1]], LabCbrtTab_b[vfX##n.val[2]], LabCbrtTab_b[vfX##n.val[3]]); \
+            vfX##n = v_int32x4(LabCbrtTab_b[v_extract_n<0>(vfX##n)], LabCbrtTab_b[v_extract_n<1>(vfX##n)], \
+                               LabCbrtTab_b[v_extract_n<2>(vfX##n)], LabCbrtTab_b[v_extract_n<3>(vfX##n)]); \
             \
             /* int fY = LabCbrtTab_b[CV_DESCALE(R*C3 + G*C4 + B*C5, lab_shift)]; */ \
             v_int32x4 vfY##n = v_fma(vR##n, v_setall_s32(C3), v_setall_s32(one_lsh_by_lab_shift_sub_one)); \
             vfY##n = v_fma(vG##n, v_setall_s32(C4), vfY##n);\
             vfY##n = v_fma(vB##n, v_setall_s32(C5), vfY##n);\
             vfY##n = v_shr<lab_shift>(vfY##n);\
-            vfY##n = v_int32x4(LabCbrtTab_b[vfY##n.val[0]], LabCbrtTab_b[vfY##n.val[1]], LabCbrtTab_b[vfY##n.val[2]], LabCbrtTab_b[vfY##n.val[3]]);\
+            vfY##n = v_int32x4(LabCbrtTab_b[v_extract_n<0>(vfY##n)], LabCbrtTab_b[v_extract_n<1>(vfY##n)], \
+                               LabCbrtTab_b[v_extract_n<2>(vfY##n)], LabCbrtTab_b[v_extract_n<3>(vfY##n)]);\
             \
             /* int fZ = LabCbrtTab_b[CV_DESCALE(R*C6 + G*C7 + B*C8, lab_shift)];*/ \
             v_int32x4 vfZ##n = v_fma(vR##n, v_setall_s32(C6), v_setall_s32(one_lsh_by_lab_shift_sub_one));\
             vfZ##n = v_fma(vG##n, v_setall_s32(C7), vfZ##n);\
             vfZ##n = v_fma(vB##n, v_setall_s32(C8), vfZ##n);\
             vfZ##n = v_shr<lab_shift>(vfZ##n);\
-            vfZ##n = v_int32x4(LabCbrtTab_b[vfZ##n.val[0]], LabCbrtTab_b[vfZ##n.val[1]], LabCbrtTab_b[vfZ##n.val[2]], LabCbrtTab_b[vfZ##n.val[3]]);\
+            vfZ##n = v_int32x4(LabCbrtTab_b[v_extract_n<0>(vfZ##n)], LabCbrtTab_b[v_extract_n<1>(vfZ##n)], \
+                               LabCbrtTab_b[v_extract_n<2>(vfZ##n)], LabCbrtTab_b[v_extract_n<3>(vfZ##n)]);\
             \
             /* int L = CV_DESCALE( Lscale*fY + Lshift, lab_shift2 );*/ \
             v_int32x4 vL##n = v_fma(vfY##n, v_setall_s32(Lscale), v_setall_s32(Lshift+one_lsh_by_lab_shift2_sub_one));\
@@ -1645,6 +1651,7 @@ struct RGB2Lab_b
             /* int b = CV_DESCALE( 200*(fY - fZ) + 128*(1 << lab_shift2), lab_shift2 );*/ \
             v_int32x4 vb##n = v_fma(vfY##n - vfZ##n, v_setall_s32(200), v_setall_s32(one_shift_lab_mul_128+one_lsh_by_lab_shift2_sub_one));\
             vb##n = v_shr<lab_shift2>(vb##n);
+
 
             // Do 4 batches of 4 RGB2Labs
             RGB2LAB_BATCH(0)