A Neon specific implementation of RGB2Lab increases single threaded performance by ~25%, here's the numbers run on aws c6gd.4xlarge with gcc 9.3 (numbers are similar using gcc 10)

| Test set   | Test number |    After/before ratio | Speedup with 1/million bounds [%] |
|------------|-------------|----------|-----------------------------------|
| cvtColor8u |           8 |  0.76835 |                        23.2 ± 0.2 |
| cvtColor8u |          34 |  0.76204 |                        23.8 ± 0.2 |
| cvtColor8u |          67 |  0.76667 |                        23.3 ± 0.2 |
| cvtColor8u |          69 |  0.76773 |                        23.2 ± 0.2 |
| cvtColor8u |          71 |  0.76231 |                        23.8 ± 0.2 |
| cvtColor8u |          73 |  0.76184 |                        23.8 ± 0.2 |
| cvtColor8u |          90 |  0.76851 |                        23.1 ± 0.2 |
| cvtColor8u |         103 |  0.76143 |                        23.9 ± 0.2 |
| cvtColor8u |         128 |  0.73870 |                        26.1 ± 0.1 |
| cvtColor8u |         154 |  0.73760 |                        26.2 ± 0.2 |
| cvtColor8u |         187 |  0.73891 |                        26.1 ± 0.1 |
| cvtColor8u |         189 |  0.73889 |                        26.1 ± 0.1 |
| cvtColor8u |         191 |  0.73802 |                        26.2 ± 0.2 |
| cvtColor8u |         193 |  0.73817 |                        26.2 ± 0.2 |
| cvtColor8u |         210 |  0.73879 |                        26.1 ± 0.1 |
| cvtColor8u |         223 |  0.73745 |                        26.3 ± 0.2 |
| cvtColor8u |         248 |  0.73756 |                        26.2 ± 0.1 |
| cvtColor8u |         274 |  0.73613 |                        26.4 ± 0.1 |
| cvtColor8u |         307 |  0.73768 |                        26.2 ± 0.1 |
| cvtColor8u |         309 |  0.73767 |                        26.2 ± 0.1 |
| cvtColor8u |         311 |  0.73676 |                        26.3 ± 0.2 |
| cvtColor8u |         313 |  0.73672 |                        26.3 ± 0.2 |
| cvtColor8u |         330 |  0.73748 |                        26.3 ± 0.1 |
| cvtColor8u |         343 |  0.73591 |                        26.4 ± 0.1 |

### Pull Request Readiness Checklist

See details at https://github.com/opencv/opencv/wiki/How_to_contribute#making-a-good-pull-request

- [x] I agree to contribute to the project under Apache 2 License.
- [x] To the best of my knowledge, the proposed patch is not based on a code under GPL or other license that is incompatible with OpenCV
- [x] The PR is proposed to proper branch
- [ ] There is reference to original bug report and related work
- [ ] There is accuracy test, performance test and test data in opencv_extra repository, if applicable
      Patch to opencv_extra has the same branch name.
- [ ] The feature is well documented and sample code can be built with the project CMake

```
force_builders=linux,docs,ARMv8,ARMv7
```