@alalek 2021-04-09
Marking this **RFC**, because it doesn't follow OpenCV guidelines to avoid using of raw native intrinsics in OpenCV modules.