@jondea 2021-04-09
Hi @alalek, thank you for looking into this. We (me and @fpetrogalli) submitted it like this because we weren't sure what correct approach was. Also, sorry if this is an silly question, but what does it mean to mark it as RFC?

One possible solution to the raw intrinsics is to keep the `#if CV_NEON` block in color_lab.cpp but rewrite it using the HAL intrinsics. Another solution would be to split it into a neon specific file, like in the case of `resize.cpp`, `resize.avx2.cpp` and `resize.sse4_1.cpp` for example. Are either of these acceptable or preferable? Or is there a another way which would achieve the same goal?