@jondea 2021-04-27
The changes have been rewritten to use just HAL intrinsics, any feedback would be appreciated.