@fpetrogalli 2021-05-25
> @jondea, thank you very much! I tested the code both on Mac-Intel and Mac-ARM (M1), it works well, the claimed acceleration is achieved. On Intel it's no slower than the previous version, but, unfortunately, it's 128-bit only.

@vpisarev , @jondea  is working on an equivalent version that uses SVE2 intrinsics. He is using the intrinsic `svld1uh_gather_s32index_s32` for the variation of `v_lut()` that does a gather from the indexes. It is a Vector Length Agnositc (VLA) version, so it could be ported easily to HAL once we make the `nlanes` field to be a runtime value and we have the correspondent indexed `lut` intrinsic.